+++ components/command/group/wallet.tsx
@@ -5,7 +5,7 @@ import { Wallet as WalletIcon, Copy, SwitchCamera, LogOut } from 'lucide-react';
 import { CommandItemComponent } from '@/components/command/index';
 import { useCommandMenu } from '@/components/command/command-menu-context';
 import { CommandGroup } from '@/components/ui/command';
-import { useToast } from '@/hooks/useToast';
+import { useToast } from '@/lib/useToast';
 
 const walletSubPages = ['wallet-list', 'wallet-connected'];
